Peticiones Retrofit y Adapter Funcionando

+++ app/src/main/java/armando/alvarez/examenibm/data/model/ListPrice.kt
@@ -5,7 +5,7 @@ import com.google.gson.annotations.SerializedName
 
 data class ListPrice(
     @SerializedName("amount")
-    var amount: Int?,
+    var amount: Double?,
     @SerializedName("currencyCode")
     var currencyCode: String?
 )
